mqtt_subscriber_app/pyproject.toml
[tool.poetry]
name = "mqtt-subscriber-app"
version = "0.1.0"
description = ""
authors = ["Sheena <11192150+sheenacodes@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
pymongo = "^4.6.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
subscribe = "mqtt_subscriber_app.subscribe:main"
